feat(crates/tuono): feat: add spinner `tuono build`
