+++ .editorconfig
@@ -5,3 +5,6 @@ end_of_line = lf
 indent_style = space
 indent_size = 2
 max_line_length = 80
+
+[*.rs]
+indent_size = 4
