+++ crates/tuono/src/cli.rs
@@ -1,11 +1,9 @@
-use fs_extra::dir::{copy, CopyOptions};
 use std::path::PathBuf;
-use std::thread::sleep;
-use std::time::Duration;
 
 use clap::{Parser, Subcommand};
 
 use crate::app::App;
+use crate::build;
 use crate::mode::Mode;
 use crate::scaffold_project;
 use crate::source_builder::{bundle_axum_source, check_tuono_folder, create_client_entry_files};
@@ -77,78 +75,9 @@ pub fn app() -> std::io::Result<()> {
             watch::watch().unwrap();
         }
         Actions::Build { ssg, no_js_emit } => {
-            let mut app = init_tuono_folder(Mode::Prod)?;
+            let app = init_tuono_folder(Mode::Prod)?;
 
-            if no_js_emit {
-                println!("Rust build successfully finished");
-                return Ok(());
-            }
-
-            if ssg && app.has_dynamic_routes() {
-                // TODO: allow dynamic routes static generation
-                println!("Cannot statically build dynamic routes");
-                return Ok(());
-            }
-
-            app.build_tuono_config()
-                .expect("Failed to build tuono.config.ts");
-
-            app.check_server_availability(Mode::Prod);
-
-            app.build_react_prod();
-
-            if ssg {
-                println!("SSG: generation started");
-
-                let static_dir = PathBuf::from("out/static");
-
-                if static_dir.is_dir() {
-                    std::fs::remove_dir_all(&static_dir)
-                        .expect("Failed to clear the out/static folder");
-                }
-
-                std::fs::create_dir(&static_dir).expect("Failed to create static output dir");
-
-                copy(
-                    "./out/client",
-                    static_dir,
-                    &CopyOptions::new().overwrite(true).content_only(true),
-                )
-                .expect("Failed to clone assets into static output folder");
-
-                // the process is killed below so we don't really need to use wait() function
-                #[allow(clippy::zombie_processes)]
-                let mut rust_server = app.run_rust_server();
-
-                let reqwest = reqwest::blocking::Client::builder()
-                    .user_agent("")
-                    .build()
-                    .expect("Failed to build reqwest client");
-
-                // Wait for server
-                let mut is_server_ready = false;
-
-                let config = app.config.as_ref().unwrap();
-
-                while !is_server_ready {
-                    let server_url =
-                        format!("http://{}:{}", config.server.host, config.server.port);
-                    if reqwest.get(server_url).send().is_ok() {
-                        is_server_ready = true
-                    }
-                    // TODO: add maximum tries
-                    sleep(Duration::from_secs(1))
-                }
-
-                for (_, route) in app.route_map {
-                    route.save_ssg_file(&reqwest)
-                }
-
-                // Close server
-                let _ = rust_server.kill();
-            };
-
-            println!("Build successfully finished");
+            let _ = build::build(app, ssg, no_js_emit);
         }
         Actions::New {
             folder_name,
